site: wire with prediction API

+++ site/src/main/java/fr/lernejo/travelsite/records/Country.java
@@ -1,7 +1,7 @@
 package fr.lernejo.travelsite.records;
 
-public record Country(String country, float temperature) {
-    public Country(String country, float temperature) {
+public record Country(String country, double temperature) {
+    public Country(String country, double temperature) {
         this.country = country;
         this.temperature = temperature;
     }
